in HexReader 1) inplemented FirmwareCheckInfo in its own module
